// src/shared/api/@generated/model/badgeCategoryResponse.ts
/**
 * Generated by orval v8.11.0 🍺
 * Do not edit manually.
 * Gravit API Docs
 * 앱센터 16.5기 동계 프로젝트 Gravit API Docs
 * OpenAPI spec version: 1.0.0
 */
import type { BadgeResponse } from './badgeResponse.ts';

export interface BadgeCategoryResponse {
  categoryId?: number;
  categoryName: string;
  order?: number;
  categoryDescription: string;
  badgeResponses: BadgeResponse[];
}