// src/shared/api/@generated/model/badgeResponse.ts
/**
 * Generated by orval v8.11.0 🍺
 * Do not edit manually.
 * Gravit API Docs
 * 앱센터 16.5기 동계 프로젝트 Gravit API Docs
 * OpenAPI spec version: 1.0.0
 */

export interface BadgeResponse {
  badgeId?: number;
  code: string;
  name: string;
  description: string;
  order?: number;
  iconId?: number;
  earned?: boolean;
}